fix: move main component of modules to base folder for better folder structure
