+++ client/src/app/main/modules/home/home-routing.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {HomeComponent} from './components/home/home.component';
+import {HomeComponent} from './home.component';
 import { AuthGuard } from 'src/app/shared/modules/auth/guards/auth-guard.service';
 
 const homeRoutes: Routes = [
