+++ client/src/app/main/modules/home/home.component.ts
@@ -6,14 +6,14 @@ import {
   OnDestroy,
   OnInit,
 } from '@angular/core';
-import {Alignments, GridSizes, Icons, IconSize, Tabs} from '../../../../../shared/static-files/enums';
-import {User} from '../../../../../shared/interfaces/user.interface';
-import {UserStore} from '../../modules/game/stores/user.store';
-import {Game} from '../../../../../shared/interfaces/game.interface';
+import {Alignments, GridSizes, Icons, IconSize, Tabs} from '../../../shared/static-files/enums';
+import {User} from '../../../shared/interfaces/user.interface';
+import {UserStore} from './modules/game/stores/user.store';
+import {Game} from '../../../shared/interfaces/game.interface';
 import {BehaviorSubject, Subject} from 'rxjs';
-import {Rule} from '../../../../../shared/interfaces/rule.interface';
-import { GameService } from '../../modules/game/services/game.service';
-import { RuleService } from '../../modules/rule/services/rule.service';
+import {Rule} from '../../../shared/interfaces/rule.interface';
+import { GameService } from './modules/game/services/game.service';
+import { RuleService } from './modules/rule/services/rule.service';
 import { TabChangeGlobalEventEmitter } from 'src/app/shared/services/tab-change.global-event-emitter';
 import { UserService } from 'src/app/shared/services/users/user.service';
 
@@ -52,7 +52,7 @@ export class HomeComponent implements OnInit, AfterContentInit, OnDestroy {
   public ngOnInit(): void {
     this.gameService.getAll()
       .subscribe((games: Game[]) => this.games$.next(games));
-    
+
     this.ruleService.getAll()
       .subscribe((rules: Rule[]) => this.rules = rules);
