+++ client/src/app/main/modules/home/modules/game/components/overview/game-table/game-table.component.scss
