+++ client/src/app/main/modules/home/modules/game/components/overview/game-table/game-table.component.ts
@@ -1,5 +1,5 @@
 import {Component, Input} from '@angular/core';
-import {Game} from '../../../../../../../../../shared/interfaces/game.interface';
+import {Game} from '../../../../../../../../shared/interfaces/game.interface';
 import {Observable} from 'rxjs';
 
 @Component({
