+++ client/src/app/main/modules/home/modules/game/components/overview/overview.component.ts
@@ -1,7 +1,7 @@
 import {Component, EventEmitter, Input, Output} from '@angular/core';
-import {Actions, Alignments, GridSizes} from '../../../../../../../../shared/static-files/enums';
+import {Actions, Alignments, GridSizes} from '../../../../../../../shared/static-files/enums';
 import {Observable} from 'rxjs';
-import {Game} from '../../../../../../../../shared/interfaces/game.interface';
+import {Game} from '../../../../../../../shared/interfaces/game.interface';
 
 @Component({
   selector: 'ui-overview',
