+++ client/src/app/main/modules/home/modules/game/components/overview/table-action/table-action.component.scss
