+++ client/src/app/main/modules/home/modules/game/components/overview/table-action/table-action.component.ts
@@ -1,7 +1,7 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {Actions, IconColor, Icons, IconSize} from '../../../../../../../../../shared/static-files/enums';
-import {IllegalInputException} from '../../../../../../../../../shared/exceptions/IllegalInputException';
-import {UnknownCaseException} from '../../../../../../../../../shared/exceptions/UnknownCaseException';
+import {Actions, IconColor, Icons, IconSize} from '../../../../../../../../shared/static-files/enums';
+import {IllegalInputException} from '../../../../../../../../shared/exceptions/IllegalInputException';
+import {UnknownCaseException} from '../../../../../../../../shared/exceptions/UnknownCaseException';
 
 @Component({
   selector: 'ui-table-action',
@@ -26,7 +26,7 @@ export class TableActionComponent implements OnInit {
 
   public IconColor = IconColor;
   public IconSize = IconSize;
-  
+
   public ngOnInit(): void {
     if (this.disabled && this.disabled.length !== this.actions.length) {
       throw new IllegalInputException('disabled and actions should be the same length');
