+++ client/src/app/main/modules/home/modules/rule/rule.component.html
