+++ client/src/app/main/modules/home/modules/rule/rule.component.ts
@@ -1,5 +1,5 @@
 import {ChangeDetectionStrategy, Component, Input} from '@angular/core';
-import {Rule} from '../../../../../../../shared/interfaces/rule.interface';
+import {Rule} from '../../../../../shared/interfaces/rule.interface';
 
 @Component({
   selector: 'app-rule',
