+++ client/src/app/main/modules/home/modules/rule/rule.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {RuleComponent} from './components/rule/rule.component';
+import {RuleComponent} from './rule.component';
 import {MatListModule} from '@angular/material/list';
 
 @NgModule({
