+++ client/src/app/main/modules/login/components/login-form/login-form.component.scss
@@ -1,7 +1,7 @@
 #loginPage {
   height: 100%;
   overflow: auto;
-  background: url(../../../../../../../assets/images/montepulciano.jpg) left top;
+  background: url(../../../../../../assets/images/montepulciano.jpg) left top;
   background-size: cover;
   display: flex;
   flex-flow: row wrap;
