+++ client/src/app/main/modules/login/login-routing.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import { LoginComponent } from './components/login/login.component';
+import { LoginComponent } from './login.component';
 
 const appRoutes: Routes = [
   {
