+++ client/src/app/main/modules/login/login.module.ts
@@ -7,8 +7,8 @@ import {MatInputModule} from '@angular/material/input';
 import {LoginRoutingModule} from './login-routing.module';
 import {CustomTranslateModule} from '../../../shared/modules/translate/custom-translate.module';
 import { AuthModule } from 'src/app/shared/modules/auth/auth.module';
-import { LoginComponent } from './components/login/login.component';
-import { LoginFormComponent } from './components/login/login-form/login-form.component';
+import { LoginComponent } from './login.component';
+import { LoginFormComponent } from './components/login-form/login-form.component';
 
 @NgModule({
   declarations: [
@@ -23,7 +23,7 @@ import { LoginFormComponent } from './components/login/login-form/login-form.com
     FormsModule,
     MatInputModule,
     CustomTranslateModule,
-    AuthModule  
+    AuthModule
   ],
 })
 export class LoginModule { }
