+++ client/src/app/main/modules/main-content/main-content-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AuthGuard } from '../../../shared/modules/auth/guards/auth-guard.service';
-import { MainContentComponent } from './components/main-content/main-content.component';
+import { MainContentComponent } from './main-content.component';
 
 const mainContentRoutes: Routes = [
   {
