+++ client/src/app/main/modules/main-content/modules/larger-screen-content/large-screen-content.component.html
