+++ client/src/app/main/modules/main-content/modules/mobile-content/mobile-content.component.html
