+++ client/src/app/main/modules/main-content/modules/mobile-content/mobile-content.component.ts
@@ -1,11 +1,11 @@
 import { ChangeDetectionStrategy, Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { NewGameStore } from 'src/app/shared/stores/new-game.store';
-import { Icons, Tabs } from '../../../../../../../shared/static-files/enums';
+import { Icons, Tabs } from '../../../../../shared/static-files/enums';
 import { AsyncBaseComponent } from 'src/app/shared/modules/async/components/async-base-component/async-base.component';
-import { Game } from '../../../../../../../shared/interfaces/game.interface';
-import { GameService } from '../../../../../home/modules/game/services/game.service';
-import { RemoveLastAddedGameStore } from '../../../../../../../shared/stores/remove-last-added-game.store';
-import { DialogOverviewComponent } from '../../../../../../../shared/modules/add-game/components/dialog-overview/dialog-overview.component';
+import { Game } from '../../../../../shared/interfaces/game.interface';
+import { GameService } from '../../../home/modules/game/services/game.service';
+import { RemoveLastAddedGameStore } from '../../../../../shared/stores/remove-last-added-game.store';
+import { DialogOverviewComponent } from '../../../../../shared/modules/add-game/components/dialog-overview/dialog-overview.component';
 import { MatDialog } from '@angular/material/dialog';
 import { TabChangeGlobalEventEmitter } from 'src/app/shared/services/tab-change.global-event-emitter';
