+++ client/src/app/main/modules/main-content/modules/mobile-content/mobile-content.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ActionBarFooterComponent } from './components/action-bar-footer/action-bar-footer.component';
-import { MobileContentComponent } from './components/mobile-content/mobile-content.component';
+import { MobileContentComponent } from './mobile-content.component';
 import { SimpleTitleBarComponent } from './components/simple-title-bar/simple-title-bar.component';
 import { CustomTranslateModule } from '../../../../../shared/modules/translate/custom-translate.module';
 import { TransitionModule } from '../../../../../shared/modules/transition/transition.module';
